Fix flutter analyze warnings
