+++ lib/service/storage_service.dart
@@ -1,6 +1,7 @@
 import 'dart:typed_data';
 
 import 'package:supabase_flutter/supabase_flutter.dart';
+// ignore: depend_on_referenced_packages
 import "package:http/http.dart" as http;
 
 class StorageService {
